Moving adapters to new folder #23
