+++ src/app/shared/adapters/reservation-fire.adapter.spec.ts
